Change component names
